mcp-host/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "mcp>=1.22.0",
    "python-dotenv>=1.2.1",
    "rich>=13.0.0",
]
